requirements.txt
numpy
python-igraph
scipy
matplotlib